refactor(backend): update explorer helper imports
