+++ apps/backend/src/rhesis/backend/app/services/test_execution.py
@@ -14,10 +14,10 @@ from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models
 from rhesis.backend.app.constants import TestResultStatus
+from rhesis.backend.app.utils.user_model_utils import get_evaluation_model, get_execution_model
 from rhesis.backend.tasks.execution.executors.data import get_test_and_prompt
 from rhesis.backend.tasks.execution.executors.metrics import determine_status_from_metrics
 from rhesis.backend.tasks.execution.executors.runners import MultiTurnRunner, SingleTurnRunner
-from rhesis.backend.tasks.execution.test import get_evaluation_model, get_execution_model
 
 logger = logging.getLogger(__name__)
